Continuing with the Mass Aggregate System and Implementation
It seems that the mass aggregate system is working, however, there are some minor tweaks I wish to make so that it actually works. For example, similar to soft body physics, I will move the points towards a static point location that follows the points around, meaning that, it should move back to the origin point via a move closer function that needs to be made.